Seperate function into modules.
Added the server call to the user model.

Seperated multiline input and input to seperate components.
