+++ src/dashboard/models/user.py
@@ -181,3 +181,16 @@ def login_user(username: str) -> User:
     flask_login.login_user(user)
 
     return user
+
+
+def add_dashboard(name: str, desc: str) -> None:
+    """Adds dashboard to mongoDB.
+
+    Args:
+        name (str): Dashboard name
+        desc (str): Dashboard description
+    """
+    test_user = login_user("dashboards-page-test-user")
+    added_dashboard = Dashboard(name=name, description=desc, created=datetime.now())
+    test_user.dashboards.append(added_dashboard)
+    test_user.save()
